update: move auth to own module
